update some api

+++ app/Http/Controllers/CategoryController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Dish;
 use App\Models\Menu;
+use App\Models\Image;
 use App\Models\Category;
+use App\Models\ImageDish;
 use Illuminate\Http\Request;
 
 class CategoryController extends Controller
@@ -14,16 +16,16 @@ class CategoryController extends Controller
         $categoryBuilder = Category::where('name', 'like', '%' . $request->name . '%');
         $categories = $categoryBuilder->get();
 
-        return ['categories' => $categories];
+        return $categories;
     }
 
     public function findOne(string $id, Request $request)
     {
         $category = Category::where('categoryId', '=', $id)->first();
         if (!$category) {
-            return ['message' => 'category does not exist'];
+            abort(404, 'category does not exist');
         }
-        return ['category' => $category];
+        return $category;
     }
 
     public function create(Request $request)
@@ -35,15 +37,20 @@ class CategoryController extends Controller
         $menu = Menu::where('menuId', '=', $request->menuId)->first();
 
         if ($menu == null) {
-            return ['message' => 'menu does not exist'];
+            abort(404, 'menu does not exist');
         }
 
-        $category = Category::create([
+        $category = ($request->categoryId == null) ? [
             'name' => $request->name,
             'menuId' => $menu->menuId
-        ]);
+        ] : [
+            'categoryId' => $request->categoryId,
+            'name' => $request->name,
+            'menuId' => $menu->menuId
+        ];
+        Category::create($category);
 
-        return ['category' => $category];
+        return "success";
     }
 
     public function update(string $id, Request $request)
@@ -54,17 +61,28 @@ class CategoryController extends Controller
         $categoryBuilder = Category::where('categoryId', '=', $id);
         $category = $categoryBuilder->first();
 
+        if (!$category) {
+            abort(404, 'category does not exist');
+        }
+
         $category->update(['name' => $request->newName]);
 
-        return ['category' => $category];
+        return "success";
     }
 
     public function delete(string $id)
     {
         $category = Category::where('categoryId', '=', $id)->first();
+        if (!$category) {
+            abort(404, 'category does not exist');
+        }
         $category->delete();
         $dishBuilder = Dish::where('categoryId', '=', $id);
+        $imageDishBuilder = ImageDish::whereIn('dishId', $dishBuilder->pluck('dishId'));
+        $imageBuilder = Image::whereIn('imageId', $imageDishBuilder->pluck('imageId'));
+
+        $imageBuilder->delete();
         $dishBuilder->delete();
-        return ['category' => $category];
+        return "success";
     }
 }
