+++ app/Http/Controllers/DishController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Dish;
+use App\Models\Image;
+use App\Models\Review;
 use App\Models\Category;
+use App\Models\ImageDish;
 use Illuminate\Http\Request;
 use App\Models\ReviewDishCrossRef;
-use App\Models\Image;
 
 class DishController extends Controller
 {
@@ -17,20 +19,18 @@ class DishController extends Controller
         $arrDishImg = [];
         foreach ($dishes as $dish) {
             $images = Image::whereIn('imageId', $dish->imageDish()->pluck('imageId'))->first();
-            $dish['images'] = $images;
+            $dish['imageDish'] = $images;
             $arrDishImg[] = $dish;
         }
-        return [
-            'dishes' => $arrDishImg
-        ];
+        return $arrDishImg;
     }
 
     public function findOne(string $id) {
         $dish = Dish::find($id);
         if (!$dish) {
-            return ['message' => 'dish does not exist'];
+            abort(404, 'dish does not exist');
         }
-        return ['dish' => $dish];
+        return $dish;
     }
 
     public function create(Request $request)
@@ -45,18 +45,25 @@ class DishController extends Controller
 
         $category = Category::where('categoryId', '=', $request->categoryId)->first();
         if (!$category) {
-            return ['message' => 'category does not exist'];
+            abort(404, 'category does not exist');
         }
-
-        $dish = Dish::create([
+        $dish = ($request->dishId) ? [
+            'dishId' => $request->dishId,
             'name' => $request->name,
             'description' => $request->description,
             'cost' => $request->cost,
             'numberOfTimesCalled' => $request->numberOfTimesCalled,
             'categoryId' => $request->categoryId
-        ]);
+        ] : [
+            'name' => $request->name,
+            'description' => $request->description,
+            'cost' => $request->cost,
+            'numberOfTimesCalled' => $request->numberOfTimesCalled,
+            'categoryId' => $request->categoryId
+        ];
+        Dish::create($dish);
 
-        return ['dish' => $dish];
+        return "success";
     }
 
     public function update(string $id, Request $request)
@@ -71,7 +78,7 @@ class DishController extends Controller
         if ($request->categoryId) {
             $category = Category::where('categoryId', '=', $request->categoryId)->first();
             if (!$category) {
-                return ['message' => 'category does not exist'];
+                abort(404, 'category does not exist');
             }
         }
 
@@ -85,21 +92,26 @@ class DishController extends Controller
                 'categoryId' => $request->categoryId ?? $dish->categoryId,
             ]);
 
-        return ['dish' => $dish];
+        return "success";
     }
 
     public function delete(string $id)
     {
+        $imageDishBuilder = ImageDish::where('dishId', '=', $id);
+        $imageBuilder = Image::whereIn('imageId', $imageDishBuilder->pluck('imageId'));
         $reviewDishCrossRefBuilder = ReviewDishCrossRef::where('dishId', '=', $id);
-        $reviewDishCrossRef = $reviewDishCrossRefBuilder->get();
+        $reviewBuilder = Review::whereIn('reviewId', $reviewDishCrossRefBuilder->pluck('reviewId'));
+
 
         $dish = Dish::where('dishId', '=', $id)->first();
-        $dish->delete();
-        $reviewDishCrossRefBuilder->delete();
+        if (!$dish) {
+            abort(404, 'dish does not exist');
+        }
 
-        return [
-            'dish' => $dish,
-            'reviewDishCrossRef' => $reviewDishCrossRef
-        ];
+        $dish->delete();
+        $reviewBuilder->delete();
+        $imageBuilder->delete();
+
+        return "success";
     }
 }
